attached_assets/requirements_1751324455355.txt
openpyxl
natsort
pyodm
opencv-python>=4.1.1
Pillow
PyYAML>=5.3.1
requests
scipy
tqdm
protobuf
django==4.1
gunicorn
psycopg2-binary
rasterio
rio_tiler
gdal==3.4.1
pandas
seaborn
fontawesomefree==6.2.1
ipython
nltk
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
